提交

+++ xgb-boot/boot-dao/src/main/java/com/xgb/dao/SysDeptMapper.java
@@ -2,12 +2,11 @@ package com.xgb.dao;
 
 import com.xgb.model.SysDept;
 import com.xgb.model.SysDeptExample;
+import java.util.List;
 import org.apache.ibatis.annotations.Param;
 
-import java.util.List;
-
 public interface SysDeptMapper {
-    long countByExample(SysDeptExample example);
+    int countByExample(SysDeptExample example);
 
     int deleteByExample(SysDeptExample example);
